+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/elements/MenuFilterable.java
@@ -20,7 +20,7 @@ public abstract class MenuFilterable extends MenuBackground {
     protected List<MenuStockItem> inventory = new ArrayList<>();
 
     protected MenuFilterBar filterBar;
-    protected MenuScrollBar scrollBar;
+    protected MenuScrollBar scrollBar, scrollBar2;
     public ResourceLocation selectedRL = null;
     public ItemStack selectedItemStack;
     int itemsX = 100, itemsY = 100, itemWidth = 140, itemHeight = 10;
@@ -41,6 +41,8 @@ public abstract class MenuFilterable extends MenuBackground {
         	filterBar.render(gui, mouseX, mouseY, partialTicks);
         if(scrollBar != null)
             scrollBar.render(gui, mouseX, mouseY, partialTicks);
+        if(scrollBar2 != null)
+            scrollBar2.render(gui, mouseX, mouseY, partialTicks);
         if (selectedItemStack == null)
             selectedItemStack = new ItemStack(BuiltInRegistries.ITEM.get(selectedRL));
 	    if (!ItemStack.matches(selectedItemStack, ItemStack.EMPTY)) {